requirements.txt
torch
torchvision
numpy
opencv-python
scikit-image
tqdm
Pillow
matplotlib
plotly
dash
pandas
patsy
statsmodels
FastGeodis
